requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
accelerate=1.0.1=pypi_0
aiohappyeyeballs=2.4.4=py311h06a4308_0
aiohttp=3.11.11=pypi_0
aiosignal=1.3.2=pypi_0
albucore=0.0.23=pypi_0
albumentations=1.4.24=pypi_0
anls-star=0.0.11=pypi_0
annotated-types=0.7.0=pypi_0
anthropic=0.45.2=pypi_0
anyio=4.6.2.post1=pypi_0
arrow-cpp=19.0.0=h865e1df_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=24.3.0=py311h06a4308_0
audioread=3.0.1=pypi_0
av=14.0.1=pypi_0
aws-c-auth=0.6.19=h5eee18b_0
aws-c-cal=0.5.20=hdbd6064_0
aws-c-common=0.8.5=h5eee18b_0
aws-c-compression=0.2.16=h5eee18b_0
aws-c-event-stream=0.2.15=h6a678d5_0
aws-c-http=0.6.25=h5eee18b_0
aws-c-io=0.13.10=h5eee18b_0
aws-c-mqtt=0.7.13=h5eee18b_0
aws-c-s3=0.1.51=hdbd6064_0
aws-c-sdkutils=0.1.6=h5eee18b_0
aws-checksums=0.1.13=h5eee18b_0
aws-crt-cpp=0.18.16=h6a678d5_0
aws-sdk-cpp=1.11.212=hecad206_0
bitsandbytes=0.44.1=pypi_0
blas=1.0=openblas
boost-cpp=1.82.0=hdb19cb5_2
bottleneck=1.4.2=py311hf4808d0_0
brotli-python=1.0.9=py311h6a678d5_9
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.12.31=h06a4308_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.0=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
coloredlogs=15.0.1=pypi_0
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.1=pypi_0
cycler=0.12.1=pypi_0
datasets=3.2.0=pypi_0
debugpy=1.8.10=py311hfdbb021_0
decorator=5.1.1=pyhd3eb1b0_0
decord=0.6.0=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.8=py311h06a4308_0
distro=1.9.0=pypi_0
doclayout-yolo=0.0.2=pypi_0
docstring-parser=0.16=pypi_0
einops=0.8.0=pypi_0
einx=0.3.0=pypi_0
encodec=0.1.1=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.2.0=pypi_0
filelock=3.16.1=pypi_0
flash-attn=2.7.4.post1=pypi_0
flatbuffers=24.3.25=pypi_0
fonttools=4.55.3=pypi_0
frozendict=2.4.6=pypi_0
frozenlist=1.5.0=py311h5eee18b_0
fsspec=2024.9.0=pypi_0
gflags=2.2.2=h6a678d5_1
gliner=0.2.14=pypi_0
glog=0.5.0=h6a678d5_1
google-ai-generativelanguage=0.6.10=pypi_0
google-api-core=2.24.0=pypi_0
google-api-python-client=2.157.0=pypi_0
google-auth=2.37.0=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-cloud-aiplatform=1.71.1=pypi_0
google-cloud-bigquery=3.29.0=pypi_0
google-cloud-core=2.4.1=pypi_0
google-cloud-resource-manager=1.14.0=pypi_0
google-cloud-storage=2.19.0=pypi_0
google-crc32c=1.6.0=pypi_0
google-generativeai=0.8.3=pypi_0
google-resumable-media=2.7.2=pypi_0
googleapis-common-protos=1.66.0=pypi_0
grpc-google-iam-v1=0.14.0=pypi_0
grpcio=1.69.0=pypi_0
grpcio-status=1.69.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.6=pypi_0
httplib2=0.22.0=pypi_0
httpx=0.27.2=pypi_0
huggingface-hub=0.26.0=pypi_0
huggingface_hub=0.24.6=py311h06a4308_0
humanfriendly=10.0=pypi_0
icecream=2.1.4=pypi_0
icu=73.1=h6a678d5_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=pyha770c72_1
ipykernel=6.29.5=pyh3099207_0
ipython=8.27.0=py311h06a4308_0
jedi=0.19.2=py311h06a4308_0
jinja2=3.1.4=pypi_0
jiter=0.6.1=pypi_0
joblib=1.4.2=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
kiwisolver=1.4.8=pypi_0
krb5=1.21.3=h143b758_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libabseil=20240116.2=cxx17_h6a678d5_0
libboost=1.82.0=h109eef0_2
libbrotlicommon=1.0.9=h5eee18b_9
libbrotlidec=1.0.9=h5eee18b_9
libbrotlienc=1.0.9=h5eee18b_9
libcurl=8.4.0=hca28451_0
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_1
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=14.2.0=h77fa898_1
libgrpc=1.62.2=h2d74bed_0
libnghttp2=1.57.0=h2d74bed_0
libopenblas=0.3.21=h043d6bf_0
libprotobuf=4.25.3=he621ea3_0
librosa=0.10.2.post1=pypi_0
libsodium=1.0.20=h4ab18f5_0
libssh2=1.11.1=h251f7ec_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=11.2.0=h1234567_1
libthrift=0.15.0=h1795dd8_2
libuuid=1.41.5=h5eee18b_0
libzlib=1.2.13=h4ab18f5_6
llvmlite=0.44.0=pypi_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=3.0.1=pypi_0
matplotlib=3.10.0=pypi_0
matplotlib-inline=0.1.6=py311h06a4308_0
mpmath=1.3.0=pypi_0
msgpack=1.1.0=pypi_0
multidict=6.1.0=py311h5eee18b_0
multiprocess=0.70.16=pypi_0
munkres=1.1.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.2.1=pypi_0
nltk=3.9.1=pypi_0
numba=0.61.0=pypi_0
numexpr=2.10.1=py311hd28fd6d_0
numpy=1.26.4=py311h24aa872_0
numpy-base=1.26.4=py311hbfb1bba_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
ollama=0.4.5=pypi_0
onnxruntime=1.19.2=pypi_0
onnxruntime-gpu=1.20.1=pypi_0
openai=1.52.0=pypi_0
opencv-python=4.10.0.84=pypi_0
opencv-python-headless=4.10.0.84=pypi_0
openssl=3.4.0=hb9d3cd8_0
orc=2.0.1=h2d29ad5_0
packaging=24.1=pypi_0
pandas=2.2.3=py311h6a678d5_0
parso=0.8.4=py311h06a4308_0
pexpect=4.8.0=pyhd3eb1b0_3
pillow=11.0.0=pypi_0
pip=24.2=py311h06a4308_0
platformdirs=4.3.6=pyhd8ed1ab_1
pooch=1.8.2=pypi_0
prompt-toolkit=3.0.43=py311h06a4308_0
prompt_toolkit=3.0.43=hd3eb1b0_0
propcache=0.2.1=pypi_0
proto-plus=1.25.0=pypi_0
protobuf=5.28.2=pypi_0
psutil=6.1.0=py311h9ecbd09_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=18.1.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pycparser=2.22=pypi_0
pydantic=2.9.2=pypi_0
pydantic-core=2.23.4=pypi_0
pygments=2.15.1=py311h06a4308_1
pyparsing=3.2.1=pypi_0
pysocks=1.7.1=py311h06a4308_0
pytesseract=0.3.13=pypi_0
python=3.11.11=he870216_0
python-dateutil=2.9.0.post0=pyhff2d567_1
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=3.5.0=py311h5eee18b_0
python_abi=3.11=2_cp311
pytz=2024.2=pypi_0
pyyaml=6.0.2=py311h5eee18b_0
pyzmq=26.2.0=py311h7deb3e3_3
qwen-vl-utils=0.0.8=pypi_0
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
regex=2024.9.11=pypi_0
requests=2.32.3=py311h06a4308_1
rsa=4.9=pypi_0
s2n=1.3.27=hdbd6064_0
safetensors=0.4.5=py311h448de89_1
scikit-learn=1.5.2=pypi_0
scipy=1.13.1=pypi_0
seaborn=0.13.2=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=75.1.0=py311h06a4308_0
shapely=2.0.7=pypi_0
simsimd=6.2.1=pypi_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.2.1=h6a678d5_0
sniffio=1.3.1=pypi_0
soundfile=0.13.1=pypi_0
soxr=0.5.0.post1=pypi_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
stringzilla=3.11.3=pypi_0
supervision=0.25.1=pypi_0
sympy=1.13.1=pypi_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=3.5.0=pypi_0
tiktoken=0.8.0=pypi_0
timm=1.0.12=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.21.0=py311h182c674_0
torch=2.5.1=pypi_0
torchaudio=2.5.1=pypi_0
torchvision=0.20.1=pypi_0
tornado=6.4.2=py311h9ecbd09_0
tqdm=4.66.5=py311h92b7b1e_0
traitlets=5.14.3=py311h06a4308_0
transformers=4.49.0.dev0=pypi_0
triton=3.1.0=pypi_0
typing-extensions=4.12.2=py311h06a4308_0
typing_extensions=4.12.2=py311h06a4308_0
tzdata=2024.2=pypi_0
uritemplate=4.1.1=pypi_0
urllib3=2.2.3=pypi_0
utf8proc=2.6.1=h5eee18b_1
vector-quantize-pytorch=1.21.4=pypi_0
verovio=4.5.1=pypi_0
vertexai=1.71.1=pypi_0
vocos=0.1.0=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.44.0=py311h06a4308_0
xxhash=0.8.0=h7f8727e_3
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.18.3=pypi_0
zeromq=4.3.5=h3b0a872_7
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=h4ab18f5_6
zstd=1.5.6=hc292b87_0
